+++ tomcat/src/main/java/org/apache/catalina/Manager.java
@@ -2,6 +2,7 @@ package org.apache.catalina;
 
 import java.io.IOException;
 import java.util.Optional;
+import org.apache.catalina.session.Session;
 
 /**
  * A <b>Manager</b> manages the pool of Sessions that are associated with a
