+++ tomcat/src/main/java/org/apache/catalina/session/SessionManager.java
@@ -1,9 +1,10 @@
-package org.apache.catalina;
+package org.apache.catalina.session;
 
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
+import org.apache.catalina.Manager;
 
 public class SessionManager implements Manager {
